requirements.txt
fastapi
uvicorn[standard]

pydantic
pydantic-settings

langchain
langchain-openai
langchain-google-genai
langchain-community
langchain-pinecone 

faiss-cpu         
pinecone-client   

pypdf             
python-docx       
unstructured      

python-dotenv

requests